начала писать тесты
к черту полетел код (точнее функция buildnode)
