+++ SkipListTest/test.cpp
@@ -1,0 +1,7 @@
+#include "pch.h"
+#include "main.c" 
+
+TEST(TestCaseName, TestName) {
+  EXPECT_EQ(1, 1);
+  EXPECT_TRUE(true);
+}
