+++ SkipListTest2/test.cpp
@@ -1,0 +1,12 @@
+#include "pch.h"
+#include "../skiplist/skiplist.h"
+#include "../skiplist/skiplist.c"
+
+TEST(TestCaseName, TestName) {
+  EXPECT_EQ(1, 1);
+  EXPECT_TRUE(true);
+}
+
+TEST(TestCaseName, SimpleTest) {
+  EXPECT_EQ(sum(1,2), 3);
+}
