+++ skiplist/Test1/test.c
@@ -1,0 +1,9 @@
+#include "pch.h"
+//#include <gtest/gtest.h>
+#include "skiplist.h"
+
+
+TEST(sum, addTwoAndFive) {
+	EXPECT_EQ(7, 7);
+}
+
